Create game_logic
